Add BaseColumnsController for column_types controllers
