+++ app/controllers/repository_columns/status_columns_controller.rb
@@ -1,9 +1,8 @@
 # frozen_string_literal: true
 
 module RepositoryColumns
-  class StatusColumnsController < ApplicationController
+  class StatusColumnsController < BaseColumnsController
     include InputSanitizeHelper
-    before_action :load_repository
     before_action :load_column, only: %i(update destroy)
     before_action :check_create_permissions, only: :create
     before_action :check_manage_permissions, only: %i(update destroy)
@@ -91,24 +90,6 @@ module RepositoryColumns
 
     def update_repository_column_params
       params.require(:repository_column).permit(repository_status_items_attributes: %i(id status icon _destroy))
-    end
-
-    def load_repository
-      @repository = Repository.accessible_by_teams(current_team).find_by(id: params[:repository_id])
-      render_404 unless @repository
-    end
-
-    def load_column
-      @repository_column = @repository.repository_columns.find_by(id: params[:id])
-      render_404 unless @repository_column
-    end
-
-    def check_create_permissions
-      render_403 unless can_create_repository_columns?(@repository)
-    end
-
-    def check_manage_permissions
-      render_403 unless can_manage_repository_column?(@repository_column)
     end
   end
 end
